requirements.txt
pandas>=2.0
numpy>=1.24
matplotlib>=3.7
scipy>=1.10
seaborn>=0.12
pyyaml>=6.0
openseespy>=3.7
